Settings page updated
