--- src/constants/storage.js
@@ -1,6 +1,0 @@
-export const ORIENTATION_SETTING_KEY = '@orientation_setting';
-export const OrientationTypes = {
-  FREE: 'free',
-  PORTRAIT: 'portrait',
-  LANDSCAPE: 'landscape',
-};
